+++ src/unicodedb/types.nim
@@ -1,0 +1,9 @@
+
+type
+  UnicodeTypeMask = enum
+    utmDecimal = 0x01
+    utmDigit = 0x02
+    utmNumeric = 0x04
+    utmLowercase = 0x08
+    utmUppercase = 0x10
+    utmCased = 0x20